test: refactor ApiClient
