+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/api/ApiClientSpec.kt
@@ -3,9 +3,6 @@ package com.rakuten.tech.mobile.miniapp.api
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.mock
 import com.rakuten.tech.mobile.miniapp.*
-import com.rakuten.tech.mobile.miniapp.TEST_MA_DISPLAY_NAME
-import com.rakuten.tech.mobile.miniapp.TEST_MA_ICON
-import com.rakuten.tech.mobile.miniapp.TEST_MA_ID
 import com.rakuten.tech.mobile.sdkutils.AppInfo
 import junit.framework.TestCase
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -119,7 +116,7 @@ open class ApiClientSpec {
         val mockCall: Call<List<MiniAppInfo>> = mock()
 
         When calling mockAppInfoApi.fetchInfo(any(), any(), any()) itReturns mockCall
-        When calling mockRequestExecutor.executeRequest(mockCall) itReturns emptyList<MiniAppInfo>()
+        When calling mockRequestExecutor.executeRequest(mockCall) itReturns emptyList()
 
         val apiClient = createApiClient(appInfoApi = mockAppInfoApi)
 
@@ -156,7 +153,7 @@ open class ApiClientSpec {
 
     @Test(expected = MiniAppSdkException::class)
     fun `should throw exception when there is authentication errors`() = runBlockingTest {
-        val executor = Mockito.spy(mockRequestExecutor)
+        val executor = spyRetrofitExecutor()
         val request: Call<String> = object : Call<String> {
             override fun enqueue(callback: Callback<String>) {}
 
@@ -177,7 +174,7 @@ open class ApiClientSpec {
 
     @Test(expected = MiniAppSdkException::class)
     fun `should throw exception when there is standard errors`() = runBlockingTest {
-        val executor = Mockito.spy(mockRequestExecutor)
+        val executor = spyRetrofitExecutor()
         val request: Call<String> = object : Call<String> {
             override fun enqueue(callback: Callback<String>) {}
 
@@ -216,6 +213,15 @@ open class ApiClientSpec {
         )
     }
 
+    @Test
+    fun `MiniAppHttpException should provide proper message`() {
+        val exception = MiniAppHttpException(
+            Response.error(400, TEST_ERROR_MSG.toResponseBody()), TEST_ERROR_MSG)
+
+        exception.message() shouldContain TEST_ERROR_MSG
+        exception.errorMessage shouldBe TEST_ERROR_MSG
+    }
+
     private fun createApiClient(
         retrofit: Retrofit = mockRetrofitClient,
         hostAppId: String = TEST_HA_ID_APP,
@@ -233,158 +239,14 @@ open class ApiClientSpec {
         manifestApi = manifestApi,
         downloadApi = downloadApi
     )
-}
 
-open class RetrofitRequestExecutorSpec private constructor(
-    internal val mockServer: MockWebServer
-) : MockWebServerBaseTest(mockServer) {
-
-    constructor() : this(MockWebServer())
-
-    private lateinit var retrofit: Retrofit
-    private lateinit var baseUrl: String
-
-    @Before
-    fun baseSetup() {
-        baseUrl = mockServer.url("/").toString()
-
-        retrofit = Retrofit.Builder()
-            .addConverterFactory(GsonConverterFactory.create())
-            .baseUrl(baseUrl)
-            .build()
+    private fun spyRetrofitExecutor(): RetrofitRequestExecutor {
+        AppInfo.instance = mock()
+        val retrofit = createRetrofitClient(
+            baseUrl = TEST_URL_HTTPS_2,
+            rasAppId = TEST_HA_ID_APP,
+            subscriptionKey = TEST_HA_SUBSCRIPTION_KEY
+        )
+        return Mockito.spy(RetrofitRequestExecutor(retrofit))
     }
-
-    internal fun createApi() = retrofit.create(TestApi::class.java)
-
-    internal fun createRequestExecutor(
-        retrofit: Retrofit = this.retrofit
-    ) = RetrofitRequestExecutor(
-        retrofit = retrofit
-    )
-}
-
-@ExperimentalCoroutinesApi
-open class RetrofitRequestExecutorNormalSpec : RetrofitRequestExecutorSpec() {
-
-    @Test
-    fun `should return the body`() = runBlockingTest {
-        mockServer.enqueue(createTestApiResponse(testValue = TEST_VALUE))
-
-        val response = createRequestExecutor()
-            .executeRequest(createApi().fetch())
-
-        response.testKey shouldEqual TEST_VALUE
-    }
-}
-
-@ExperimentalCoroutinesApi
-open class RetrofitRequestExecutorErrorSpec : RetrofitRequestExecutorSpec() {
-
-    @Test(expected = MiniAppSdkException::class)
-    fun `should throw when server returns error response`() = runBlockingTest {
-        mockServer.enqueue(createErrorResponse())
-        createRequestExecutor()
-            .executeRequest(createApi().fetch())
-    }
-
-    @Test(expected = MiniAppSdkException::class)
-    fun `should throw exception when the response returned by server is not of type T`() =
-        runBlockingTest {
-            mockServer.enqueue(createInvalidTestApiResponse())
-            createRequestExecutor()
-                .executeRequest(createApi().fetch())
-        }
-
-    @Test
-    fun `should throw exception with the error message returned by server`() = runBlockingTest {
-        mockServer.enqueue(createErrorResponse(message = TEST_ERROR_MSG))
-
-        try {
-            createRequestExecutor()
-                .executeRequest(createApi().fetch())
-
-            TestCase.fail("Should have thrown ErrorResponseException.")
-        } catch (exception: MiniAppSdkException) {
-            exception.message.toString() shouldContain TEST_ERROR_MSG
-        }
-    }
-
-    @Test
-    fun `should append error message to base exception message`() = runBlockingTest {
-        mockServer.enqueue(createErrorResponse(message = TEST_ERROR_MSG))
-
-        try {
-            createRequestExecutor()
-                .executeRequest(createApi().fetch())
-
-            TestCase.fail("Should have thrown ErrorResponseException.")
-        } catch (exception: MiniAppSdkException) {
-            exception.message.toString() shouldContain TEST_ERROR_MSG
-        }
-    }
-
-    @Test
-    fun `should throw exception with the error message returned by server for 401 & 403`() =
-        runBlockingTest {
-            mockServer.enqueue(createAuthErrorResponse(message = TEST_ERROR_MSG))
-            try {
-                createRequestExecutor()
-                    .executeRequest(createApi().fetch())
-
-                TestCase.fail("Should have thrown ErrorResponseException.")
-            } catch (exception: MiniAppSdkException) {
-                exception.message.toString() shouldContain TEST_ERROR_MSG
-            }
-        }
-
-    @Test
-    fun `should append default message when server doesn't return error message`() =
-        runBlockingTest {
-            mockServer.enqueue(
-                MockResponse()
-                    .setResponseCode(400)
-                    .setBody("{}")
-            )
-
-            try {
-                createRequestExecutor()
-                    .executeRequest(createApi().fetch())
-
-                TestCase.fail("Should have thrown ErrorResponseException.")
-            } catch (exception: MiniAppSdkException) {
-                exception.message.toString() shouldContain "No error message"
-            }
-        }
-
-    private val standardErrorBody = { code: Int, message: String ->
-        """
-            {
-                "code": $code,
-                "message": "$message"
-            }
-        """.trimIndent()
-    }
-
-    private val authErrorBody = { code: String, message: String ->
-        """
-            {
-                "code": "$code",
-                "message": "$message"
-            }
-        """.trimIndent()
-    }
-
-    private fun createErrorResponse(
-        code: Int = 400,
-        message: String = TEST_ERROR_MSG
-    ): MockResponse = MockResponse()
-        .setResponseCode(code)
-        .setBody(standardErrorBody(code, message))
-
-    private fun createAuthErrorResponse(
-        code: Int = 401,
-        message: String = TEST_ERROR_MSG
-    ) = MockResponse()
-        .setResponseCode(code)
-        .setBody(authErrorBody(code.toString(), message))
 }
