+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/api/RetrofitRequestExecutorSpec.kt
@@ -1,0 +1,133 @@
+package com.rakuten.tech.mobile.miniapp.api
+
+import com.rakuten.tech.mobile.miniapp.MiniAppSdkException
+import com.rakuten.tech.mobile.miniapp.TEST_ERROR_MSG
+import com.rakuten.tech.mobile.miniapp.TEST_VALUE
+import junit.framework.TestCase
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.test.runBlockingTest
+import okhttp3.mockwebserver.MockResponse
+import okhttp3.mockwebserver.MockWebServer
+import org.amshove.kluent.shouldContain
+import org.amshove.kluent.shouldEqual
+import org.junit.Before
+import org.junit.Test
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+open class RetrofitRequestExecutorSpec private constructor(
+    internal val mockServer: MockWebServer
+) : MockWebServerBaseTest(mockServer) {
+
+    constructor() : this(MockWebServer())
+
+    private lateinit var retrofit: Retrofit
+    private lateinit var baseUrl: String
+
+    @Before
+    fun baseSetup() {
+        baseUrl = mockServer.url("/").toString()
+
+        retrofit = Retrofit.Builder()
+            .addConverterFactory(GsonConverterFactory.create())
+            .baseUrl(baseUrl)
+            .build()
+    }
+
+    internal fun createApi() = retrofit.create(TestApi::class.java)
+
+    internal fun createRequestExecutor(
+        retrofit: Retrofit = this.retrofit
+    ) = RetrofitRequestExecutor(
+        retrofit = retrofit
+    )
+}
+
+@ExperimentalCoroutinesApi
+open class RetrofitRequestExecutorNormalSpec : RetrofitRequestExecutorSpec() {
+
+    @Test
+    fun `should return the body`() = runBlockingTest {
+        mockServer.enqueue(createTestApiResponse(testValue = TEST_VALUE))
+
+        val response = createRequestExecutor()
+            .executeRequest(createApi().fetch())
+
+        response.testKey shouldEqual TEST_VALUE
+    }
+}
+
+@ExperimentalCoroutinesApi
+open class RetrofitRequestExecutorErrorSpec : RetrofitRequestExecutorSpec() {
+
+    @Test(expected = MiniAppSdkException::class)
+    fun `should throw exception when the response returned by server is not of type T`() =
+        runBlockingTest {
+            mockServer.enqueue(createInvalidTestApiResponse())
+            createRequestExecutor()
+                .executeRequest(createApi().fetch())
+        }
+
+    @Test
+    fun `should throw exception with the error message returned by server`() = runBlockingTest {
+        mockServer.enqueue(createErrorResponse(message = TEST_ERROR_MSG))
+
+        try {
+            createRequestExecutor()
+                .executeRequest(createApi().fetch())
+
+            TestCase.fail("Should have thrown ErrorResponseException.")
+        } catch (exception: MiniAppSdkException) {
+            exception.message.toString() shouldContain TEST_ERROR_MSG
+        }
+    }
+
+    @Test
+    fun `should append error message to base exception message`() = runBlockingTest {
+        mockServer.enqueue(createErrorResponse(message = TEST_ERROR_MSG))
+
+        try {
+            createRequestExecutor()
+                .executeRequest(createApi().fetch())
+
+            TestCase.fail("Should have thrown ErrorResponseException.")
+        } catch (exception: MiniAppSdkException) {
+            exception.message.toString() shouldContain TEST_ERROR_MSG
+        }
+    }
+
+    @Test
+    fun `should append default message when server doesn't return error message`() =
+        runBlockingTest {
+            mockServer.enqueue(
+                MockResponse()
+                    .setResponseCode(400)
+                    .setBody("{}")
+            )
+
+            try {
+                createRequestExecutor()
+                    .executeRequest(createApi().fetch())
+
+                TestCase.fail("Should have thrown ErrorResponseException.")
+            } catch (exception: MiniAppSdkException) {
+                exception.message.toString() shouldContain "No error message"
+            }
+        }
+
+    private val standardErrorBody = { code: Int, message: String ->
+        """
+            {
+                "code": $code,
+                "message": "$message"
+            }
+        """.trimIndent()
+    }
+
+    private fun createErrorResponse(
+        code: Int = 400,
+        message: String = TEST_ERROR_MSG
+    ): MockResponse = MockResponse()
+        .setResponseCode(code)
+        .setBody(standardErrorBody(code, message))
+}
